server/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
dnspython==2.6.1
fastapi==0.115.0
google-auth==2.35.0
idna==3.10
motor==3.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.9.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
